+++ src/algorithms/RandomCliffordCircuit.cpp
@@ -206,12 +206,13 @@ auto append2QClifford(QuantumComputation& circ, const std::uint16_t idx,
 
 auto createRandomCliffordCircuit(const Qubit nq, const std::size_t depth,
                                  const std::size_t seed) -> QuantumComputation {
-  auto qc = QuantumComputation(nq, nq, seed);
+  auto qc = QuantumComputation(nq, nq);
   qc.setName("random_clifford_" + std::to_string(nq) + "_" +
              std::to_string(depth) + "_" + std::to_string(seed));
 
   std::uniform_int_distribution<std::uint16_t> distribution(0, 11520);
-  auto cliffordGenerator = [&]() { return distribution(qc.getGenerator()); };
+  auto generator = std::mt19937_64(seed);
+  auto cliffordGenerator = [&]() { return distribution(generator); };
 
   for (std::size_t l = 0; l < depth; ++l) {
     if (nq == 1) {